# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank REST API
  description: REST API для управления банковскими картами, пользователями и транзакциями
  version: 4.7.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер
security:
  - bearerAuth: [ ]
paths:
...
  /api/admin/users:
    post:
      summary: Создание пользователя (админ)
      description: Создание нового пользователя администратором
      operationId: createUser
      tags:
        - Администрирование
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            example:
              username: newuser
              password: password123
              roles: [ USER ]
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Получение списка пользователей (админ)
      description: Получение списка всех пользователей
      operationId: getAllUsers
      tags:
        - Администрирование
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
  /api/admin/users/{id}:
    get:
      summary: Получение пользователя по ID (админ)
      description: Получение информации о пользователе по его ID
      operationId: getUserById
      tags:
        - Администрирование
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Обновление пользователя (админ)
      description: Обновление данных пользователя администратором
      operationId: updateUser
      tags:
        - Администрирование
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Пользователь обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление пользователя (админ)
      description: Удаление пользователя по ID
      operationId: deleteUser
      tags:
        - Администрирование
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь удален
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UserRequest:
      type: object
      properties:
        username:
          type: string
          description: Имя пользователя
        password:
          type: string
          description: Пароль
        roles:
          type: array
          items:
            type: string
            enum: [ USER, ADMIN ]
          description: Роли пользователя
      required:
        - username
        - password
        - roles
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
        username:
          type: string
          description: Имя пользователя
      required:
        - id
        - username
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об ошибке
        status:
          type: integer
          description: HTTP статус код
      required:
        - message
        - status
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT